# openapi/rhythm-v1.yaml
openapi: 3.1.0
info:
  title: Rhythm API
  version: 0.0.1-current
  description: Public control-plane contract for Rhythm synthetic monitoring and deployment validation.
servers:
  - url: http://localhost:18080
    description: Local Docker
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/applications/{applicationId}/opensearch-alerts/service-assignment:
    post:
      tags: [Alerts]
      operationId: assignOpenSearchAlertsToService
      summary: Tag selected OpenSearch receiver alerts to an application service
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenSearchAlertServiceAssignmentInput"
      responses:
        "200":
          description: Alerts were assigned atomically
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/OpenSearchAlertServiceAssignmentResult"
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Error:
      description: Safe error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ResponseMeta:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string }
        page:
          type: object
          properties:
            nextCursor: { type: [string, "null"] }
            limit: { type: integer }
    OpenSearchAlertServiceAssignmentInput:
      type: object
      additionalProperties: false
      required: [alertIds, serviceId]
      properties:
        alertIds:
          type: array
          minItems: 1
          maxItems: 200
          uniqueItems: true
          items: { type: string, format: uuid }
        serviceId:
          description: Application service UUID, or an empty string to clear the assignment.
          anyOf:
            - type: string
              format: uuid
            - const: ""
    OpenSearchAlertServiceAssignmentResult:
      type: object
      required: [assignedCount]
      properties:
        assignedCount: { type: integer, minimum: 0, maximum: 200 }
        serviceId: { type: string, format: uuid }
        serviceName: { type: string }
...